Move `PromptStore` to `prompt_library` crate
