+++ crates/assistant/Cargo.toml
@@ -43,7 +43,6 @@ fuzzy.workspace = true
 globset.workspace = true
 gpui.workspace = true
 handlebars.workspace = true
-heed.workspace = true
 html_to_markdown.workspace = true
 http_client.workspace = true
 indexed_docs.workspace = true
@@ -63,6 +62,7 @@ parking_lot.workspace = true
 paths.workspace = true
 picker.workspace = true
 project.workspace = true
+prompt_library.workspace = true
 proto.workspace = true
 regex.workspace = true
 release_channel.workspace = true
