+++ crates/assistant/src/slash_command/default_command.rs
@@ -1,4 +1,3 @@
-use crate::prompt_library::PromptStore;
 use anyhow::{anyhow, Result};
 use assistant_slash_command::{
     ArgumentCompletion, SlashCommand, SlashCommandOutput, SlashCommandOutputSection,
@@ -6,6 +5,7 @@ use assistant_slash_command::{
 };
 use gpui::{Task, WeakView};
 use language::{BufferSnapshot, LspAdapterDelegate};
+use prompt_library::PromptStore;
 use std::{
     fmt::Write,
     sync::{atomic::AtomicBool, Arc},
