+++ crates/prompt_library/Cargo.toml
@@ -1,0 +1,28 @@
+[package]
+name = "prompt_library"
+version = "0.1.0"
+edition = "2021"
+publish = false
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+path = "src/prompt_library.rs"
+
+[dependencies]
+anyhow.workspace = true
+chrono.workspace = true
+collections.workspace = true
+futures.workspace = true
+fuzzy.workspace = true
+gpui.workspace = true
+heed.workspace = true
+parking_lot.workspace = true
+paths.workspace = true
+rope.workspace = true
+serde.workspace = true
+text.workspace = true
+util.workspace = true
+uuid.workspace = true
