+++ crates/prompt_library/src/prompt_library.rs
@@ -1,0 +1,9 @@
+mod prompt_store;
+
+use gpui::AppContext;
+
+pub use crate::prompt_store::*;
+
+pub fn init(cx: &mut AppContext) {
+    prompt_store::init(cx);
+}
